moved write amplify meta to its own file
