+++ packages/amplify-category-notifications/commands/notifications/configure.js
@@ -1,6 +1,7 @@
 const inquirer = require('inquirer');
 const pinpointHelper = require('../../lib/pinpoint-helper');
 const notificationManager = require('../../lib/notifications-manager');
+const writeAmplifyMeta = require('../../lib/writeAmplifyMeta');
 
 module.exports = {
   name: 'configure',
@@ -21,10 +22,10 @@ module.exports = {
       channelName = answer.selection;
     }
 
-    await pinpointHelper.checkPinpointApp(context);
-    notificationManager.updateaServiceMeta(context);
+    await pinpointHelper.ensurePinpointApp(context);
+    writeAmplifyMeta(context);
     await notificationManager.configureChannel(context, channelName);
-    notificationManager.updateaServiceMeta(context);
+    writeAmplifyMeta(context);
 
     return context;
   },
