+++ packages/amplify-category-notifications/commands/notifications/remove.js
@@ -1,6 +1,7 @@
 const inquirer = require('inquirer');
 const pinpointHelper = require('../../lib/pinpoint-helper');
 const notificationManager = require('../../lib/notifications-manager');
+const writeAmplifyMeta = require('../../lib/writeAmplifyMeta');
 
 const PinpointApp = 'The Pinpoint application';
 const Cancel = 'Cancel';
@@ -34,10 +35,10 @@ module.exports = {
 
     if (channelName && channelName !== Cancel) {
       if (channelName !== PinpointApp) {
-        await pinpointHelper.checkPinpointApp(context);
-        notificationManager.updateaServiceMeta(context);
+        await pinpointHelper.ensurePinpointApp(context);
+        writeAmplifyMeta(context);
         await notificationManager.disableChannel(context, channelName);
-        notificationManager.updateaServiceMeta(context);
+        writeAmplifyMeta(context);
       } else if (pinpointHelper.isAnalyticsAdded(context)) {
         context.print.error('Execution aborted.');
         context.print.info('You have an analytics resource in your backend tied to the Amazon Pinpoint resource');
@@ -52,7 +53,7 @@ module.exports = {
         if (answer.deletePinpointApp) {
           await pinpointHelper.deletePinpointApp(context);
           context.print.info('The Pinpoint application has been successfully deleted.');
-          notificationManager.updateaServiceMeta(context);
+          writeAmplifyMeta(context);
         }
       }
     }
