+++ packages/amplify-category-notifications/lib/notifications-manager.js
@@ -75,29 +75,11 @@ async function configureChannel(context, channelName) {
   }
 }
 
-function updateaServiceMeta(context) {
-  const amplifyMetaFilePath = context.amplify.pathManager.getAmplifyMetaFilePath();
-  if (context.exeInfo && context.exeInfo.serviceMeta) {
-    context.exeInfo.serviceMeta.lastPushTimeStamp = new Date();
-  }
-  let jsonString = JSON.stringify(context.exeInfo.amplifyMeta, null, '\t');
-  fs.writeFileSync(amplifyMetaFilePath, jsonString, 'utf8');
-
-  const currentAmplifyMetaFilePath = context.amplify.pathManager.getCurentAmplifyMetaFilePath();
-  const currentAmplifyMeta = JSON.parse(fs.readFileSync(currentAmplifyMetaFilePath));
-  currentAmplifyMeta[constants.CategoryName] = context.exeInfo.amplifyMeta[constants.CategoryName];
-  jsonString = JSON.stringify(currentAmplifyMeta, null, '\t');
-  fs.writeFileSync(currentAmplifyMetaFilePath, jsonString, 'utf8');
-
-  context.amplify.onCategoryOutputsChange(context);
-}
-
 module.exports = {
   getAvailableChannels,
   getEnabledChannels,
   getDisabledChannels,
   enableChannel,
   disableChannel,
-  configureChannel,
-  updateaServiceMeta,
+  configureChannel
 };
